sDES 1.0
Okay that was surprisingly easy.
Basically decryption is THE EXACT SAME PROCESS as encryption, with the exception of course, that the keys are introduced in inverse order.
To solve this I created a huge ass function to encrypt, which receives the text to encrypt, and the keys. To use that function for decryption, the input for the text should be the ciphertext, and the two original keys in inverse order.